finished pingbox module with CBA settings
